BLING, TS, "not a number"s
the plot will skip any "not a number"s instead of using a line to connect breaks.  I have changed some of the BLING calls since I needed those changes to get them working.  Also, I have put the entire signal calculation into "if total signal is checked" option because total signal was being calculated even when we just wanted BLING.  It may be better to put it all into a module that we import so it's less messy.