changes to BLINGs(particularly CMB)
for CMB, I have commented out some of the math and added calculations that look more like the equation but the plot is wrong so this could have been a mistake.  In these changes, I assumed v0 is the exactly the frequency used in the paper's equation.  I'm not sure where the "(3e10)/((2 * resol * 0.1))" used to calculate i_a and i_b come from so that could be the issue.